HAL: return pretained preview buffer to surface when preview window changed or camera disconnected.
When preview is paused due to taking snapshot, preview buffer is retained.
However, if UI changes preview window or disconnect, previous preview buffer is
not valid any more, so we need to put it back. Next time when preview is
started, it will act as a clean start of preview and get preview buffer
from display.

Change-Id: Ia3c63e44c3e6d004d3b347a2a3c7d312e2aacdb2
CRs-fixed: 342926

Conflicts:
	QCameraHWI.cpp
	QCameraHWI_Preview.cpp