camera: use same ion client for buffer cache operations.
In the new kernel, every new ion device that is opened
creates a new client, irrespective of whether it's from the same
process or not. Hence, the buffer cache operations will
have to be carried out on the same fd that was
used to allocate the buffers. This change makes
the necessary adjustments to achieve the same.

Change-Id:
Ic4e245f9f8f2fbe2606251cb97837cfdc01a72a3

This fixed cache invalidation errors